Simplify ClusterAdd
This moves ClusterAdd away from using the presentational & container
component approach. I didn't touch ClusterLogin as it's a much more
complex component.
